requirements.txt
streamlit
streamlit-option-menu
textblob
text2emotion
plotly
pandas
pillow
opencv-python
numpy
requests
nltk
flair
fer
torch
torchvision
